readme

+++ script-deploy.sh
@@ -1,7 +1,10 @@
  #!/bin/bash
 clear
 
-sudo rm -rf practicas-empresa-ull
+if [ -d "practicas-empresa-ull" ]; then
+  sudo rm -rf practicas-empresa-ull
+fi
+
 git clone https://github.com/kaizten/practicas-empresa-ull
 cd practicas-empresa-ull
 cd kaizten-blue-task
@@ -13,8 +16,11 @@ echo 'Kaizten Red Task installed'
 
 cd ..
 
-sudo rm -rf $HOME/docker/volumes/mongodb/*
-mkdir $HOME/docker/volumes/mongodb/
+if [ -d "$HOME/docker/volumes/mongodb/" ]; then
+  sudo rm -rf $HOME/docker/volumes/mongodb/*
+else
+  mkdir $HOME/docker/volumes/mongodb/
+fi
 
 export DATAFLOW_VERSION=2.4.2.RELEASE
 export SKIPPER_VERSION=2.3.2.RELEASE
